// BeekeepingMonitoring.ClientApp/src/app/@core/app-api.ts
export class SensorReferenceModel implements ISensorReferenceModel {
    id!: number;
    name!: string;
    description!: string | null;

    constructor(data?: ISensorReferenceModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : <any>null;
            this.name = _data["name"] !== undefined ? _data["name"] : <any>null;
            this.description = _data["description"] !== undefined ? _data["description"] : <any>null;
        }
    }

    static fromJS(data: any): SensorReferenceModel {
        data = typeof data === 'object' ? data : {};
        let result = new SensorReferenceModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        data["description"] = this.description !== undefined ? this.description : <any>null;
        return data;
    }

    clone(): SensorReferenceModel {
        const json = this.toJSON();
        let result = new SensorReferenceModel();
        result.init(json);
        return result;
    }
}

export interface ISensorReferenceModel {
    id: number;
    name: string;
    description: string | null;
}
// ...
export class CustomRuleCreateModel implements ICustomRuleCreateModel {
    sensorId!: number;
    min!: number | null;
    max!: number | null;
    programDirective!: string | null;
    ruleText!: string | null;

    constructor(data?: ICustomRuleCreateModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.sensorId = _data["sensorId"] !== undefined ? _data["sensorId"] : <any>null;
            this.min = _data["min"] !== undefined ? _data["min"] : <any>null;
            this.max = _data["max"] !== undefined ? _data["max"] : <any>null;
            this.programDirective = _data["programDirective"] !== undefined ? _data["programDirective"] : <any>null;
            this.ruleText = _data["ruleText"] !== undefined ? _data["ruleText"] : <any>null;
        }
    }

    static fromJS(data: any): CustomRuleCreateModel {
        data = typeof data === 'object' ? data : {};
        let result = new CustomRuleCreateModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["sensorId"] = this.sensorId !== undefined ? this.sensorId : <any>null;
        data["min"] = this.min !== undefined ? this.min : <any>null;
        data["max"] = this.max !== undefined ? this.max : <any>null;
        data["programDirective"] = this.programDirective !== undefined ? this.programDirective : <any>null;
        data["ruleText"] = this.ruleText !== undefined ? this.ruleText : <any>null;
        return data;
    }

    clone(): CustomRuleCreateModel {
        const json = this.toJSON();
        let result = new CustomRuleCreateModel();
        result.init(json);
        return result;
    }
}

export interface ICustomRuleCreateModel {
    sensorId: number;
    min: number | null;
    max: number | null;
    programDirective: string | null;
    ruleText: string | null;
}
// ...
export class CustomRuleUpdateModel implements ICustomRuleUpdateModel {
    sensorId!: number;
    min!: number | null;
    max!: number | null;
    programDirective!: string | null;
    ruleText!: string | null;

    constructor(data?: ICustomRuleUpdateModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.sensorId = _data["sensorId"] !== undefined ? _data["sensorId"] : <any>null;
            this.min = _data["min"] !== undefined ? _data["min"] : <any>null;
            this.max = _data["max"] !== undefined ? _data["max"] : <any>null;
            this.programDirective = _data["programDirective"] !== undefined ? _data["programDirective"] : <any>null;
            this.ruleText = _data["ruleText"] !== undefined ? _data["ruleText"] : <any>null;
        }
    }

    static fromJS(data: any): CustomRuleUpdateModel {
        data = typeof data === 'object' ? data : {};
        let result = new CustomRuleUpdateModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["sensorId"] = this.sensorId !== undefined ? this.sensorId : <any>null;
        data["min"] = this.min !== undefined ? this.min : <any>null;
        data["max"] = this.max !== undefined ? this.max : <any>null;
        data["programDirective"] = this.programDirective !== undefined ? this.programDirective : <any>null;
        data["ruleText"] = this.ruleText !== undefined ? this.ruleText : <any>null;
        return data;
    }

    clone(): CustomRuleUpdateModel {
        const json = this.toJSON();
        let result = new CustomRuleUpdateModel();
        result.init(json);
        return result;
    }
}

export interface ICustomRuleUpdateModel {
    sensorId: number;
    min: number | null;
    max: number | null;
    programDirective: string | null;
    ruleText: string | null;
}